+++ src/main/groovy/com/ullink/NuGetPack.groovy
@@ -2,15 +2,14 @@ package com.ullink
 
 import groovy.util.XmlSlurper
 import groovy.util.slurpersupport.GPathResult
-import groovy.xml.StreamingMarkupBuilder
 import groovy.xml.XmlUtil
 import org.gradle.api.tasks.InputFile
-import org.gradle.api.tasks.OutputFile;
+import org.gradle.api.tasks.OutputFile
+import org.gradle.api.Task
 
 class NuGetPack extends BaseNuGet {
 
     def nuspecFile
-    Closure nuspec
     def csprojPath
 
     def destinationDir = project.convention.plugins.base.distsDir
@@ -54,7 +53,7 @@ class NuGetPack extends BaseNuGet {
 
         if (basePath) args '-BasePath', basePath
 
-        def version = spec.metadata.version ?: project.version
+        def version = spec?.metadata?.version ?: project.version
         if (version) args '-Version', version
 
         if (exclude) args '-Exclude', exclude
@@ -71,16 +70,19 @@ class NuGetPack extends BaseNuGet {
         super.exec()
     }
 
-    void nuspec(Closure closure) {
-        nuspec = closure
-    }
-
-    Closure getNuspecCustom() {
-        nuspec
-    }
-
     GPathResult getNuspec() {
-        new XmlSlurper().parse(getNuSpecFile())
+        def file = getNuSpecFile()
+        if (file) {
+            new XmlSlurper().parse(getNuSpecFile())
+        }
+        def nugetSpec = getDependentNugetSpec()
+        if (nugetSpec) {
+            new XmlSlurper(false, false).parseText(nugetSpec.generateNuspec())
+        }
+    }
+
+    Task getDependentNugetSpec() {
+        dependsOn.find { it instanceof NuGetSpec }
     }
 
     // Because Nuget pack handle csproj or nuspec file we should be able to use it in plugin
@@ -92,59 +94,18 @@ class NuGetPack extends BaseNuGet {
     }
 
     File getNuSpecFile() {
-        if (!this.nuspecFile || !project.file(this.nuspecFile).exists()) {
-            this.nuspecFile = generateNuspecFile()
+        if (nuspecFile) {
+            project.file(this.nuspecFile)
         }
-        project.file(this.nuspecFile)
+        def nugetSpec = getDependentNugetSpec()
+        if (nugetSpec) {
+            nugetSpec.nuspecFile
+        }
     }
 
     File getPackageFile() {
         def spec = getNuspec()
         def version = spec.metadata.version ?: project.version
         new File(getDestinationDir(), spec.metadata.id.toString() + '.' + version + '.nupkg')
-    }
-
-    File generateNuspecFile() {
-        def nuspecFile = new File(temporaryDir, project.name + '.nuspec')
-        def nuspec = supplementDefaultValueOnNuspec generateNuspec()
-        nuspecFile.withWriter('UTF-8') { writer ->
-            XmlUtil.serialize (nuspec, writer)
-        }
-        nuspecFile
-    }
-
-    String generateNuspec() {
-        def builder = new StreamingMarkupBuilder()
-        builder.bind {
-            'package' (xmlns: 'http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd') {
-                if (nuspecCustom) {
-                    nuspecCustom.resolveStrategy = DELEGATE_FIRST
-                    nuspecCustom.delegate = delegate
-                    nuspecCustom.call()
-                }
-            }
-        }.toString()
-    }
-
-    GPathResult supplementDefaultValueOnNuspec(String nuspecString) {
-
-        def root = new XmlSlurper(false, false).parseText(nuspecString)
-
-        def defaultValues = {}
-        def applyDefaultValue = { String node, String value ->
-            if (root.metadata[node].isEmpty()) {
-                defaultValues <<= { delegate."$node" value }
-            }
-        }
-        applyDefaultValue ('id', project.name)
-        applyDefaultValue ('version', project.version)
-        applyDefaultValue ('description', project.description)
-
-        if (root.metadata.isEmpty()) {
-            root.appendNode { metadata defaultValues }
-        } else {
-            root.metadata.appendNode defaultValues
-        }
-        root
     }
 }