+++ src/main/groovy/com/ullink/NuGetPush.groovy
@@ -11,19 +11,17 @@ class NuGetPush extends BaseNuGet {
 
     NuGetPush() {
         super('push')
-        conventionMapping.map "nupkgFile", { project.tasks.nugetPack.packageFile }
-        project.afterEvaluate {
-            if (nupkgFile) {
-                def tasks = project.tasks.withType(NuGetPack.class).findAll { it.packageFile == nupkgFile }
-                if (tasks.size() == 1)
-                    dependsOn tasks[0]
-            }
-        }
+    }
+
+    String getNugetPackOutputFile() {
+        def tasks = project.tasks.withType(NuGetPack.class)
+        if (tasks.size() == 1)
+           tasks.first().packageFile
     }
 
     @Override
     void exec() {
-        args getNupkgFile()
+        args nupkgFile ?: nugetPackOutputFile
 
         if (serverUrl) args '-Source', serverUrl
         if (apiKey) args '-ApiKey', apiKey